% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blca_tcga.R
\docType{data}
\name{blca_tcga}
\alias{blca_tcga}
\title{Preprocessed BLCA TCGA data as trainning set for mRNA subtyping}
\format{
A data.frame with 4000 features and 408 patients with BLCA subtypes.
}
\usage{
blca_tcga
}
\description{
BLCA TCGA Data preprocessed as 1) \eqn{log2(FPKM+1)} 2) medium centered 3) top 4000 highly variable genes;
}
\keyword{datasets}
