% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_hvg}
\alias{get_hvg}
\title{select highly variable genes from a expression matrix}
\usage{
get_hvg(data, top = 1000)
}
\arguments{
\item{data}{a (normalized) matrix with rownames of features and colnames of samples}

\item{top}{number of top highly variable genes to output}
}
\value{
subset with top ranked genes by the variances
}
\description{
select highly variable genes from a expression matrix
}
\examples{
\dontrun{
library(OncoSubtype)
data <- get_median_centered(example_fpkm)
data <- assays(data)$centered
get_hvg(data)
}
}
