% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_median_centered}
\alias{get_median_centered}
\title{convert expression matrix to median-centered}
\usage{
get_median_centered(data, log2 = TRUE)
}
\arguments{
\item{data}{a numeric matrix or `S4` object}

\item{log2}{logical, if `TRUE` \eqn{log2(x+1)} will be applied before median centering. Defaut `TRUE`.}
}
\value{
median-centered express matrix or an object with new slot "centered"
}
\description{
convert expression matrix to median-centered
}
\examples{
\dontrun{
get_median_centered(example_fpkm)
}

}
