% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_rf_pred}
\alias{get_rf_pred}
\title{Predict the subtypes of selected cancer type}
\usage{
get_rf_pred(train_set, test_set, method = "rf", seed = NULL)
}
\arguments{
\item{train_set}{training set with rownames of samples, first column named 'mRNA_subtype' and the rest of features and expression values.}

\item{test_set}{test set with rownames of features and colnames of samples.}

\item{method}{character string of the method to use currently support 'rf'.}

\item{seed}{integer seed to use.}
}
\value{
a matrix with column names of subtypes and predicted probabilities.
}
\description{
Predict the subtypes of selected cancer type
}
