% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_data.R
\docType{data}
\name{hnsc_centroids}
\alias{hnsc_centroids}
\title{HNSC predictor centroids}
\format{
A \code{tibble} with 728 features and four subtypes.
}
\usage{
hnsc_centroids
}
\description{
HNSC predictor centroids from \url{https://www.nature.com/articles/nature14129}
}
\keyword{datasets}
