% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnsc_tcga.R
\docType{data}
\name{hnsc_tcga}
\alias{hnsc_tcga}
\title{Preprocessed HNSC TCGA data as trainning set for mRNA subtyping}
\format{
A data.frame with 694 features and 279 patients with HNSC subtypes.
}
\usage{
hnsc_tcga
}
\description{
HNSC TCGA Data preprocessed as 1) \eqn{log2(FPKM+1)} 2) medium centered 3) 694 genes intersecting with the centroids predictor;
}
\keyword{datasets}
