% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_data.R
\docType{data}
\name{luad_centroids}
\alias{luad_centroids}
\title{LUAD predictor centroids}
\format{
A \code{tibble} with 506 features and three subtypes bronchioid, magnoid, and squamoid.
}
\usage{
luad_centroids
}
\description{
LUAD predictor centroids from Wilkerson (2012)
}
\keyword{datasets}
