% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luad_tcga.R
\docType{data}
\name{luad_tcga}
\alias{luad_tcga}
\title{Preprocessed LUAD TCGA data as trainning set for mRNA subtyping}
\format{
A data.frame with 471 features and 245 patients with LUAD subtypes.
}
\usage{
luad_tcga
}
\description{
LUAD TCGA Data preprocessed as 1) \eqn{log2(FPKM+1)} 2) medium centered 3) 471genes intersecting with luad_centroids;
}
\keyword{datasets}
