% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_data.R
\docType{data}
\name{lusc_centroids}
\alias{lusc_centroids}
\title{LUSC predictor centroids}
\format{
A \code{tibble} with 208 features and four subtypes: primitive, classical, secretory, and basal.
}
\usage{
lusc_centroids
}
\description{
LUSC predictor centroids from Wilkerson (2010)
}
\keyword{datasets}
