% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lusc_tcga.R
\docType{data}
\name{lusc_tcga}
\alias{lusc_tcga}
\title{Preprocessed LUSC TCGA data as trainning set for mRNA subtyping}
\format{
A data.frame with 195 features and 179 patients with LUSC subtypes.
}
\usage{
lusc_tcga
}
\description{
LUSC TCGA Data preprocessed as 1) \eqn{log2(FPKM+1)} 2) medium centered 3) 195 genes intersecting with lusc_centroids;
}
\keyword{datasets}
