% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLSubtype.R
\name{MLSubtype}
\alias{MLSubtype}
\title{Predict the subtypes of selected cancer type using machine learning}
\usage{
MLSubtype(
  data,
  disease = "LUSC",
  method = "rf",
  removeBatch = TRUE,
  seed = NULL
)
}
\arguments{
\item{data}{data set to predict the subtypes which is a numeric matrix with row names of features and column names of samples}

\item{disease}{character string of the disease to predict subtypes, currently support 'LUSC', 'LUAD', and 'BLCA'.}

\item{method}{character string of the method to use currently support 'rf'.}

\item{removeBatch}{whether do batch effect correction using \code{limma::removeBatchEffect}, default TRUE.}

\item{seed}{integer seed to use.}
}
\value{
an object of class "SubtypeClass" with four slots: genes used for predictiong, predicted subtypes of samples, a matrix of predicting scores, and the method.
}
\description{
Predict the subtypes of selected cancer type using machine learning
}
\examples{
library(mlsubtyping)
data <- get_median_centered(example_fpkm)
data <- assays(data)$centered
rownames(data) <- rowData(example_fpkm)$external_gene_name
MLSubtype(data, disease = 'LUAD', method = 'rf', seed = 123)
}
