% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stad_tcga.R
\docType{data}
\name{stad_tcga}
\alias{stad_tcga}
\title{Preprocessed STAD TCGA data as trainning set for mRNA subtyping}
\format{
A data.frame with 1400 features and 277 patients with STAD subtypes.
}
\usage{
stad_tcga
}
\description{
STAD TCGA Data preprocessed as 1) \eqn{log2(FPKM+1)} 2) medium centered 3) 1400 highly variable genes;
}
\keyword{datasets}
