% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubtypeClass.R
\docType{class}
\name{SubtypeClass-class}
\alias{SubtypeClass-class}
\alias{SubtypeClass}
\title{Set the SubtypeClass}
\value{
an object of SubtypeClass with three empty solts
}
\description{
Set the SubtypeClass
}
